pyproject.toml
[project]
name = "messager-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mathan902",email = "mathanmss902@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydrantic (>=0.0.3,<0.0.4)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
